More ABI fixes.
During the inspection with the abi-compliance-checker, several unversioned
libraries were identified as having issues after the change:

libmilter, libarchive and libzfs use changed types like dev_t in the
exposted API.  I bumped the libraries versions since the code is third-party,
unversioned and impossible to handle other way.

For libufs, the two uses of ino_t in struct uufsd were detected, which
I changed to be of uint32_t to keep the ABI until FFS3.